Keep helper methods in util
